--- src/components/WorkWithUs.tsx/WorkWithUs.tsx
@@ -1,17 +1,0 @@
-const WorkWithUs = () => {
-    return (
-        <section className="flex-col justify-center align-middle content-center">
-            <h2 className="flex justify-start font-bebas text-xs text-white mx-[1.5rem] my-[1rem]">
-                Work With Us
-            </h2>
-            <span className="flex justify-start font-oxanium text-xss text-white mx-[1.625rem] my-4 leading-6">
-            Trust us with everything about your website. We'll make sure you get the style and quality you’re looking for to represent your Company. 
-            </span>
-            <span className="flex justify-start font-oxanium text-xss text-white mx-[1.625rem] leading-6">
-            Send us a message with your contact information and we’ll reach out to you at our earliest convenience.
-            </span>
-        </section>
-    );
-};
-
-export default WorkWithUs;
