+++ src/components/WorkWithUs/WorkWithUs.tsx
@@ -1,0 +1,17 @@
+const WorkWithUs = () => {
+    return (
+        <section className="flex-col justify-center align-middle content-center">
+            <h2 className="flex justify-start font-bebas text-xs text-white mx-[1.5rem] my-[1rem] md:text-[2.5rem] md:mx-[3.125rem] xl:text-[3.125rem] xl:mx-[4.25rem] xl:my-16 ">
+                Work With Us
+            </h2>
+            <span className="flex justify-start font-oxanium text-xss text-white mx-[1.625rem] my-4 leading-6 md:text-[1.25rem] md:leading-[2.5rem] md:mx-[3.125rem] md:my-8 xl:text-[1.5rem] xl:mx-[4.25rem] xl:my-12 ">
+            Trust us with everything about your website. We'll make sure you get the style and quality you’re looking for to represent your Company. 
+            </span>
+            <span className="flex justify-start font-oxanium text-xss text-white mx-[1.625rem] leading-6 md:text-[1.25rem] md:leading-[2.5rem] md:mx-[3.125rem] md:my-8 xl:text-[1.5rem] xl:mx-[4.25rem] xl:my-12 ">
+            Send us a message with your contact information and we’ll reach out to you at our earliest convenience.
+            </span>
+        </section>
+    );
+};
+
+export default WorkWithUs;
